# services/api-gateway/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Customer Support Ticketing System API
  version: 1.0.0
  description: >
    Public API surface exposed via the API Gateway. Auth endpoints are
    unauthenticated; all others require a Bearer access token issued by
    POST /api/auth/login.
servers:
  - url: /api
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 10 }
                firstName: { type: string }
                lastName: { type: string }
                role: { type: string, enum: [CUSTOMER, AGENT, MANAGER, ADMIN], default: CUSTOMER }
      responses:
        '201': { description: User registered }
        '409': { description: Email already taken }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in and receive access/refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: Login successful, returns accessToken/refreshToken }
        '401': { description: Invalid credentials }
...
  /tickets:
    post:
      tags: [Tickets]
      summary: Create a ticket (customer)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject, description, categoryId]
              properties:
                subject: { type: string }
                description: { type: string }
                categoryId: { type: integer }
                priority: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
      responses:
        '201': { description: Ticket created }
    get:
      tags: [Tickets]
      summary: List tickets, scoped to the caller's role
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Paginated ticket list }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT