# services/ticket-service/src/app.js
'use strict';

require('express-async-errors');
const express = require('express');
const helmet = require('helmet');
const cors = require('cors');
const morgan = require('morgan');
const rateLimit = require('express-rate-limit');

const config = require('./config/env');
const logger = require('./utils/logger');
const traceId = require('./middleware/traceId');
const ticketRoutes = require('./routes/ticketRoutes');
const { errorHandler, notFoundHandler } = require('./middleware/errorHandler');

const app = express();

app.use(helmet());
app.use(cors({ origin: process.env.CORS_ORIGIN || '*', credentials: true }));
app.use(express.json({ limit: '200kb' })); // ticket descriptions can be sizeable
app.use(traceId);
app.use(morgan('combined', { stream: { write: (msg) => logger.info(msg.trim()) } }));

app.use(rateLimit({
  windowMs: config.rateLimit.windowMs,
  max: config.rateLimit.max,
  standardHeaders: true,
  legacyHeaders: false,
}));

app.get('/health', (req, res) => res.status(200).json({ status: 'UP', service: config.serviceName }));
app.get('/ready', async (req, res) => {
  try {
    const { sequelize } = require('./config/database');
    await sequelize.authenticate();
    res.status(200).json({ status: 'READY' });
  } catch (err) {
    res.status(503).json({ status: 'NOT_READY', reason: err.message });
  }
});

app.use('/api/tickets', ticketRoutes);

// Serve uploaded attachments statically behind auth would normally go
// through a signed-URL download endpoint rather than a static mount;
// left as a documented follow-up rather than exposing /uploads directly.

app.use(notFoundHandler);
app.use(errorHandler);

module.exports = app;